fixes of version 1 connector after code review
