+++ Connector/artifacts/contracts/MyNFT.sol/MyNFT.dbg.json
@@ -1,0 +1,4 @@
+{
+  "_format": "hh-sol-dbg-1",
+  "buildInfo": "..\\..\\build-info\\5dba171e91429f44ef3023227557fb44.json"
+}
